backend/tools/gmail_email_tool/pyproject.toml
[project]
name = "gmail-email-tool"
version = "0.1.0"
description = "Standalone Gmail API email sender with OAuth, retries, logging, and idempotency."
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.165.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
]

[project.scripts]
gmail-email-tool = "gmail_email_tool.cli:main"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
